pyproject.toml
[project]
name = "seclang_parser"
dynamic = [ "version" ]
description = "SecLang Parser using ANTLR"
authors = [
  {name = "Felipe Zipitria", email = "felipe.zipitria@owasp.org"}
]
requires-python = ">=3.11"
license = "Apache-2.0"
readme = "README.md"
keywords = ["OWASP", "CRS", "SecLang", "ModSecurity", "Coraza"]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "antlr4-python3-runtime >=4.13.2",
]

[project.scripts]
  seclang_parser = 'seclang_parser.cli:run'

[project.urls]
  issues = "https://github.com/coreruleset/seclang_parser/issues"
  homepage = "https://github.com/coreruleset/seclang_parser"
  repository = "https://github.com/coreruleset/seclang_parser.git"

[dependency-groups]
  dev = [
    "pytest >=8.1.1,<9",
    "pyyaml >=6.0"
  ]

[build-system]
  requires = ["hatchling", "hatch-vcs"]
  build-backend = "hatchling.build"

[tool.hatch.version]
  source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "no-guess-dev"

[[tool.uv.index]]
  name = "pypi"
  url = "https://pypi.org/simple/"
  publish-url = "https://pypi.org/legacy/"

[[tool.uv.index]]
  name = "testpypi"
  url = "https://test.pypi.org/simple/"
  publish-url = "https://test.pypi.org/legacy/"
